Add SQL to pair with Full.py
This will work with the full-refresh Python script, but could probably
use some tidying up since the idea of always having to copy the entire
thing from the loader_fct to the price_paid_fct is a little annoying.
That said, COPY should work much more quickly than INSERT, and having a
check for duplicates in the loader_fct has its advantages.