Changing from tables to materialised views
The materialised view approach should be better for recreating
aggregate tables on-the-fly as it saves copying a lot of data around
for no real purpose. The one thing I need to check is how this
behaviour works going forward when all we want to do is update the
Price Paid data set with the latest monthly dump.