Extending view creation
Added more views. Opted not to create views at postcode level as the
volume of transactions is too low. Still having issues with the Python
script and how to figure out a) where we are, and b) where we need to
be, in order to load the .dbconfig file and other components
successfully.