Bug fixes (utils, imports)
Improved way that code tries to find out where it’s being run from and
chdir() to the right place so that it can import the utils module and
get access to the SQL and other scripts that it needs. Also fixed a
weird bug in utils.py that was mucking the interpolation up (terrible
practice to try to dynamically rewrite a dictionary anyway).